added message reading
theres now a place where you can read messages. the nations list has
order status on the left and the chat button is blue if that nation sent
you a message and you havent read it yet. mail icon changed to blue
circle because it looks nicer. other message-centric additions