added notifs
should check every ~half hour and show a notification if you have no
orders in a game or if you have a new message in a game. tapping the
notif should open an associated gameactivity. should also fire even when
the app has been removed from background (swiped away in recent apps)

not sure if any of this is actually true but it seems to work ok