fixed gameactivity bug
the app was not able to get proper access to the user's webdippy account
if they had come back to the app after a while (credential validation
and cookie handling had been done exclusively when the mainactivity
started). now the game activity also logs in, so there should be fewer
problems with that.